requirements.txt
beautifulsoup4
requests
pandas
numpy>=1.19.0
scikit-learn
sentence-transformers==2.2.2
fastapi
uvicorn
streamlit
faiss-cpu==1.7.4
google-genai>=1.0.0
python-dotenv>=0.19.0
markdown>=3.0.0
flask
